Avoid deadlock from pymongo after celery worker forks
Mongo clients use threads (when a replicaset is in use) and celery
workers use fork.  This combination is known to be unsafe, as threads
from a parent don't survive a fork, and might leave resources locked
in the new child process.

This problem is documented in
http://api.mongodb.com/python/current/faq.html#using-pymongo-with-multiprocessing
which gives the advice:

> Care must be taken when using instances of MongoClient with fork().
> Specifically, instances of MongoClient must not be copied from a
> parent process to a child process.

This change adjusts Pulp's database connection to follow that advice:
once initialized, the connection will be discarded and reinitialized in
any child process.